Added Upload Package.JSON Component
Added a basic React Component to upload dependency file - performing a check on the file Mime Type; if it is a JSON, sending an axios call to server to process it. If it is not a JSON Mime Type then I'm showing an alert, asking user to upload a valid JSON. 

Added to new dependencies 1. Axios 2. Form-Data