Receiving Package.JSON file
Receiving Package.JSON file using Multer, checking for MIME type, uploading if the type is valid. I have written and commented the code for uploading the file to Amazon S3 using knox. I'm unsure of the use-case, if it is to be uploaded to S3, please uncomment the knox part and add a config file to it. 

Questions: Is the next step to re-direct the user to (the currently named) profile page and show RecommendationList and DependencyTable? If yes, please let me know, I can work on it. This is my first pull request to an open collective repo, so I have sent very minimal amount of changes. PS: Great work!

Added dependencies: 1. Knox. 2. Multer.